chore: rename breadbook backend to BreadDiary
