+++ .gitignore
@@ -55,3 +55,4 @@ coverage/
 
 docker-compose.override.yml
 *.local.yml
+.omx/
