+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'breadbook'
+rootProject.name = 'bread-diary'
