+++ src/main/java/com/bean/breaddiary/BreadDiaryApplication.java
@@ -1,4 +1,4 @@
-package com.bean.breadbook;
+package com.bean.breaddiary;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableJpaAuditing
 @SpringBootApplication
-public class BreadbookApplication {
+public class BreadDiaryApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BreadbookApplication.class, args);
+		SpringApplication.run(BreadDiaryApplication.class, args);
 	}
 
 }
