+++ src/test/java/com/bean/breaddiary/BreadDiaryApplicationTests.java
@@ -1,10 +1,10 @@
-package com.bean.breadbook;
+package com.bean.breaddiary;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BreadbookApplicationTests {
+class BreadDiaryApplicationTests {
 
 	@Test
 	void contextLoads() {
